services/prediction/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
azure-core==1.38.3
azure-cosmos==4.15.0
azure-servicebus==7.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
isodate==0.7.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.32.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
bcrypt==4.1.2
httpx==0.27.0
